requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
click==8.1.6
fastapi==0.101.0
h11==0.14.0
idna==3.4
outcome==1.3.0.post0
pydantic==2.1.1
pydantic_core==2.4.0
PySocks==1.7.1
python-dotenv==1.0.0
selenium==4.15.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.19
starlette==0.27.0
trio==0.23.0
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==2.0.7
uvicorn==0.23.2
wsproto==1.2.0
